refactor: extract report analytics business logic from reports.v2.js
Move health scoring, AI suggestions, insights, trend predictions,
and safeCount into backend/lib/reportAnalytics/ modules.
Route file reduced from ~997 to ~610 lines.
No behavior changes — pure extraction.
