+++ backend/routes/reports.v2.js
@@ -12,406 +12,20 @@
 import { Router } from 'express';
 import { getSupabaseClient } from '../lib/supabase-db.js';
 import logger from '../lib/logger.js';
+import {
+  calculatePipelineHealth,
+  calculateLeadHealth,
+  generateDashboardSuggestions,
+  generateDashboardInsights,
+  buildDashboardAiContext,
+  warnIfSlow,
+  safeCount,
+} from '../lib/reportAnalytics/index.js';
 
-const ENABLE_AI_ENRICHMENT = process.env.AI_ENRICHMENT_ENABLED !== 'false';
-const SLOW_THRESHOLD_MS = parseInt(process.env.AI_CONTEXT_SLOW_THRESHOLD_MS || '500', 10);
 const BUNDLE_TTL_MS = 60_000; // 60 seconds cache
 
 // In-memory cache for v2 bundles
 const v2BundleCache = new Map();
-
-/**
- * Log warning if processing exceeds threshold
- */
-function warnIfSlow(operation, processingTime) {
-  if (processingTime > SLOW_THRESHOLD_MS) {
-    logger.warn(
-      `[reports.v2] SLOW: ${operation} took ${processingTime}ms (threshold: ${SLOW_THRESHOLD_MS}ms)`,
-    );
-  }
-}
-
-/**
- * Create stub AI context when enrichment is disabled or fails
- */
-function createStubAiContext(startTime, error = null) {
-  return {
-    confidence: 0,
-    suggestions: [],
-    predictions: null,
-    insights: error ? [`AI enrichment unavailable: ${error}`] : ['AI enrichment disabled'],
-    trends: null,
-    healthScore: null,
-    processingTime: Date.now() - startTime,
-    _stub: true,
-  };
-}
-
-/**
- * Calculate pipeline health based on stage distribution
- */
-function calculatePipelineHealth(stages) {
-  if (!stages || stages.length === 0) return { score: 0, status: 'no_data' };
-
-  const total = stages.reduce((sum, s) => sum + (parseInt(s.count) || 0), 0);
-  if (total === 0) return { score: 0, status: 'no_data' };
-
-  // Healthy pipeline should have good distribution across stages
-  const wonStages = stages.filter((s) =>
-    ['won', 'closed_won', 'closed'].includes(s.stage?.toLowerCase()),
-  );
-  const lostStages = stages.filter((s) => ['lost', 'closed_lost'].includes(s.stage?.toLowerCase()));
-  const activeStages = stages.filter(
-    (s) => !['won', 'closed_won', 'closed', 'lost', 'closed_lost'].includes(s.stage?.toLowerCase()),
-  );
-
-  const wonCount = wonStages.reduce((sum, s) => sum + (parseInt(s.count) || 0), 0);
-  const lostCount = lostStages.reduce((sum, s) => sum + (parseInt(s.count) || 0), 0);
-  const activeCount = activeStages.reduce((sum, s) => sum + (parseInt(s.count) || 0), 0);
-
-  // Calculate win rate for closed deals
-  const closedTotal = wonCount + lostCount;
-  const winRate = closedTotal > 0 ? wonCount / closedTotal : 0;
-
-  // Score based on: win rate (50%), active pipeline (30%), distribution (20%)
-  let score = 0;
-  score += winRate * 50; // Up to 50 points for win rate
-  score += Math.min(30, (activeCount / Math.max(total, 1)) * 100); // Up to 30 for active pipeline
-  score += activeStages.length >= 3 ? 20 : activeStages.length * 7; // Up to 20 for stage diversity
-
-  let status = 'critical';
-  if (score >= 75) status = 'healthy';
-  else if (score >= 50) status = 'needs_attention';
-  else if (score >= 25) status = 'at_risk';
-
-  return { score: Math.round(score), status, winRate: Math.round(winRate * 100) };
-}
-
-/**
- * Calculate lead conversion health
- */
-function calculateLeadHealth(statuses) {
-  if (!statuses || statuses.length === 0) return { score: 0, status: 'no_data' };
-
-  const total = statuses.reduce((sum, s) => sum + (parseInt(s.count) || 0), 0);
-  if (total === 0) return { score: 0, status: 'no_data' };
-
-  const converted = statuses.find((s) => s.status?.toLowerCase() === 'converted');
-  const qualified = statuses.find((s) => s.status?.toLowerCase() === 'qualified');
-  const contacted = statuses.find((s) => s.status?.toLowerCase() === 'contacted');
-  const newLeads = statuses.find((s) => s.status?.toLowerCase() === 'new');
-
-  const convertedCount = parseInt(converted?.count) || 0;
-  const qualifiedCount = parseInt(qualified?.count) || 0;
-  const contactedCount = parseInt(contacted?.count) || 0;
-  const newCount = parseInt(newLeads?.count) || 0;
-
-  // Conversion rate
-  const conversionRate = convertedCount / Math.max(total, 1);
-
-  // Lead progression score (are leads moving through pipeline?)
-  const progressionScore = (qualifiedCount + contactedCount) / Math.max(total - convertedCount, 1);
-
-  // Stagnation indicator (too many new, unworked leads)
-  const stagnationRisk = newCount / Math.max(total, 1);
-
-  // Calculate overall score
-  let score = 0;
-  score += conversionRate * 40; // Up to 40 for conversion
-  score += progressionScore * 35; // Up to 35 for progression
-  score += (1 - stagnationRisk) * 25; // Up to 25 for low stagnation
-  score = Math.min(100, score * 100);
-
-  let status = 'critical';
-  if (score >= 70) status = 'healthy';
-  else if (score >= 45) status = 'needs_attention';
-  else if (score >= 20) status = 'at_risk';
-
-  return {
-    score: Math.round(score),
-    status,
-    conversionRate: Math.round(conversionRate * 100),
-    stagnationRisk: Math.round(stagnationRisk * 100),
-  };
-}
-
-/**
- * Generate AI suggestions based on dashboard data
- */
-function generateDashboardSuggestions(stats, pipelineHealth, leadHealth) {
-  const suggestions = [];
-
-  // Pipeline suggestions
-  if (pipelineHealth.status === 'at_risk' || pipelineHealth.status === 'critical') {
-    suggestions.push({
-      action: 'review_stalled_opportunities',
-      priority: 'high',
-      reason: `Pipeline health is ${pipelineHealth.status} (score: ${pipelineHealth.score}/100)`,
-      confidence: 0.85,
-      category: 'pipeline',
-    });
-  }
-
-  if (pipelineHealth.winRate !== undefined && pipelineHealth.winRate < 30) {
-    suggestions.push({
-      action: 'analyze_lost_deals',
-      priority: 'high',
-      reason: `Win rate is ${pipelineHealth.winRate}% - analyze lost opportunities`,
-      confidence: 0.9,
-      category: 'pipeline',
-    });
-  }
-
-  // Lead suggestions
-  if (leadHealth.stagnationRisk > 50) {
-    suggestions.push({
-      action: 'work_new_leads',
-      priority: 'high',
-      reason: `${leadHealth.stagnationRisk}% of leads are unworked - prioritize outreach`,
-      confidence: 0.9,
-      category: 'leads',
-    });
-  }
-
-  if (leadHealth.conversionRate < 10) {
-    suggestions.push({
-      action: 'improve_qualification',
-      priority: 'medium',
-      reason: `Low conversion rate (${leadHealth.conversionRate}%) - review lead sources`,
-      confidence: 0.75,
-      category: 'leads',
-    });
-  }
-
-  // Activity suggestions
-  if (stats.activitiesLast30Days < 10) {
-    suggestions.push({
-      action: 'increase_activity',
-      priority: 'medium',
-      reason: 'Low activity volume in last 30 days',
-      confidence: 0.8,
-      category: 'activities',
-    });
-  }
-
-  // Contact/Account ratio
-  if (stats.totalContacts > 0 && stats.totalAccounts > 0) {
-    const ratio = stats.totalContacts / stats.totalAccounts;
-    if (ratio < 1.5) {
-      suggestions.push({
-        action: 'add_more_contacts',
-        priority: 'low',
-        reason: `Low contacts-per-account ratio (${ratio.toFixed(1)})`,
-        confidence: 0.65,
-        category: 'data_quality',
-      });
-    }
-  }
-
-  return suggestions;
-}
-
-/**
- * Generate insights from dashboard data
- */
-function generateDashboardInsights(stats, pipelineHealth, leadHealth) {
-  const insights = [];
-
-  // Overall data health
-  const totalRecords =
-    (stats.totalContacts || 0) +
-    (stats.totalAccounts || 0) +
-    (stats.totalLeads || 0) +
-    (stats.totalOpportunities || 0);
-
-  if (totalRecords === 0) {
-    insights.push('No CRM data found - start by adding leads or accounts');
-    return insights;
-  }
-
-  // Pipeline insights
-  if (pipelineHealth.score !== undefined) {
-    insights.push(`Pipeline health score: ${pipelineHealth.score}/100 (${pipelineHealth.status})`);
-    if (pipelineHealth.winRate !== undefined) {
-      insights.push(`Historical win rate: ${pipelineHealth.winRate}%`);
-    }
-  }
-
-  // Lead insights
-  if (leadHealth.score !== undefined) {
-    insights.push(`Lead funnel health: ${leadHealth.score}/100 (${leadHealth.status})`);
-    if (stats.newLeadsLast30Days > 0) {
-      insights.push(`${stats.newLeadsLast30Days} new leads in last 30 days`);
-    }
-  }
-
-  // Activity velocity
-  if (stats.activitiesLast30Days > 0) {
-    const dailyAvg = (stats.activitiesLast30Days / 30).toFixed(1);
-    insights.push(`Activity velocity: ${dailyAvg} activities/day`);
-  }
-
-  // Open opportunity focus
-  if (stats.openOpportunities > 0) {
-    insights.push(`${stats.openOpportunities} open opportunities require attention`);
-  }
-
-  return insights;
-}
-
-/**
- * Generate trend predictions
- */
-function generateTrendPredictions(stats, pipelineHealth, leadHealth) {
-  const predictions = {
-    nextMonth: {},
-    trends: {},
-    recommendations: [],
-  };
-
-  // Lead volume trend prediction
-  const monthlyLeadRate = stats.newLeadsLast30Days || 0;
-  predictions.nextMonth.expectedNewLeads = monthlyLeadRate; // Assume stable
-  predictions.trends.leadVolume =
-    monthlyLeadRate > 10 ? 'stable' : monthlyLeadRate > 5 ? 'low' : 'critical';
-
-  // Activity trend prediction
-  const monthlyActivityRate = stats.activitiesLast30Days || 0;
-  predictions.nextMonth.expectedActivities = monthlyActivityRate;
-  predictions.trends.activityLevel =
-    monthlyActivityRate > 30 ? 'high' : monthlyActivityRate > 10 ? 'moderate' : 'low';
-
-  // Conversion prediction based on current funnel
-  if (leadHealth.conversionRate !== undefined) {
-    predictions.nextMonth.expectedConversions = Math.round(
-      (stats.totalLeads || 0) * (leadHealth.conversionRate / 100) * 0.1,
-    );
-  }
-
-  // Win prediction based on pipeline
-  if (stats.openOpportunities > 0 && pipelineHealth.winRate !== undefined) {
-    predictions.nextMonth.expectedWins = Math.round(
-      stats.openOpportunities * (pipelineHealth.winRate / 100) * 0.15,
-    );
-  }
-
-  // Recommendations based on predictions
-  if (predictions.trends.leadVolume === 'critical') {
-    predictions.recommendations.push('Increase lead generation efforts');
-  }
-  if (predictions.trends.activityLevel === 'low') {
-    predictions.recommendations.push('Schedule more customer touchpoints');
-  }
-
-  return predictions;
-}
-
-/**
- * Build AI context for dashboard bundle
- */
-async function buildDashboardAiContext(stats, tenant_id) {
-  const startTime = Date.now();
-
-  if (!ENABLE_AI_ENRICHMENT) {
-    return createStubAiContext(startTime);
-  }
-
-  try {
-    const supabase = getSupabaseClient();
-
-    // Fetch pipeline and lead status data in parallel
-    const [pipelineResult, leadStatusResult] = await Promise.all([
-      (async () => {
-        try {
-          let q = supabase.from('v_opportunity_pipeline_by_stage').select('stage, count');
-          if (tenant_id) q = q.eq('tenant_id', tenant_id);
-          const { data } = await q;
-          return data || [];
-        } catch {
-          return [];
-        }
-      })(),
-      (async () => {
-        try {
-          let q = supabase.from('v_lead_counts_by_status').select('status, count');
-          if (tenant_id) q = q.eq('tenant_id', tenant_id);
-          const { data } = await q;
-          return data || [];
-        } catch {
-          return [];
-        }
-      })(),
-    ]);
-
-    // Calculate health scores
-    const pipelineHealth = calculatePipelineHealth(pipelineResult);
-    const leadHealth = calculateLeadHealth(leadStatusResult);
-
-    // Generate AI components
-    const suggestions = generateDashboardSuggestions(stats, pipelineHealth, leadHealth);
-    const insights = generateDashboardInsights(stats, pipelineHealth, leadHealth);
-    const predictions = generateTrendPredictions(stats, pipelineHealth, leadHealth);
-
-    // Calculate overall health score (weighted average)
-    let overallHealth = 0;
-    let weightSum = 0;
-    if (pipelineHealth.score !== undefined && pipelineHealth.status !== 'no_data') {
-      overallHealth += pipelineHealth.score * 0.5;
-      weightSum += 0.5;
-    }
-    if (leadHealth.score !== undefined && leadHealth.status !== 'no_data') {
-      overallHealth += leadHealth.score * 0.5;
-      weightSum += 0.5;
-    }
-    const healthScore = weightSum > 0 ? Math.round(overallHealth / weightSum) : null;
-
-    const processingTime = Date.now() - startTime;
-    warnIfSlow('dashboard-ai-context', processingTime);
-
-    return {
-      confidence: 0.82,
-      suggestions,
-      predictions,
-      insights,
-      trends: predictions.trends,
-      healthScore: {
-        overall: healthScore,
-        pipeline: pipelineHealth,
-        leads: leadHealth,
-      },
-      processingTime,
-    };
-  } catch (error) {
-    logger.error('[reports.v2] AI context error:', error.message);
-    return createStubAiContext(startTime, error.message);
-  }
-}
-
-/**
- * Safe count helper (copied from v1)
- */
-async function safeCount(_, table, tenant_id, filterFn, opts = {}) {
-  try {
-    const supabase = getSupabaseClient();
-    const { includeTestData = true, countMode = 'exact' } = opts;
-    let q = supabase.from(table).select('*', { count: countMode, head: true });
-    if (tenant_id) q = q.eq('tenant_id', tenant_id);
-    if (filterFn) q = filterFn(q);
-    if (!includeTestData) {
-      try {
-        q = q.or('is_test_data.is.false,is_test_data.is.null');
-      } catch {
-        /* ignore */ void 0;
-      }
-    }
-    const { count, error } = await q;
-    if (error) throw error;
-    return count ?? 0;
-  } catch (err) {
-    logger.error(`[reports.v2] safeCount error for ${table}:`, err.message);
-    return 0;
-  }
-}
 
 export default function createReportsV2Router(_pgPool) {
   const router = Router();
